Splits the tests into multiple files.
